// src/app/models/pessoa.ts
import { ContasBancaria } from './contas-bancaria';
import { Endereco } from './endereco';
import { Contato } from './contato';

export interface Pessoa {
	id?: number;
	descricao?: string;
	name?: string;
	cpfOnCnpj?: string;
	rg?: string;
	salario?: number;
	periodoFerias?: string;
	carteiraprofissional?: string;
	namewar?: string;
	status?: string;
	avatar?: string;
	avatarView?: string;
	email?: string;
	senha?: string;
	rolers?: string;
	telefones?: string[];
	instantCreation?: Date;
	dataAdmissao?: Date;
	natureza?: string;
	code?: string;
	data_situacao?: string;
	tipo?: string;
	porte?: string;
	natureza_juridica?: string;
	capital_social?: string;
	typePerson?: string;
	enderecos?: Endereco[];
	enderecoPrincipal?: Endereco;
	contasBancarias?: ContasBancaria[];
	tipoSanguineo?: string;
	atividade_principal?: string;
	nameFantasia?: string;
	contatos?: Contato[];
	contatoPrincipal?: Contato;
	telefone?: string;
	street?: string;
	zipCode?: string;
	bairro?: string;
	city?: string;
	nro?: string;
	complement?: string;
	state?: string;
	funcao?:string;

}